requirements.txt
pdfplumber
python-docx
PyYAML
gitpython
chardet
spacy
